requirements.txt
discord.py
python-dotenv
requests
sqalchemy